minor bug fixes
git-svn-id: svn://svn.code.sf.net/p/stars-nova/svn/trunk@944 03d337be-0f81-4a3b-8898-4a24ff344f0f

+++ Common/Waypoints/SplitMergeTask.cs
@@ -315,7 +315,7 @@ namespace Nova.Common.Waypoints
                     { 
                         mostVessels = moveCount;
                         designName = left.Composition[key].Design.Name;
-                        right.Name = designName + " #" + fleetNumber.ToString();  //Name of new fleet is taken from the design of the most populous vessel in the fleet
+                        right.Name = designName + " #" + ((int) fleetNumber).ToString();  //Name of new fleet is taken from the design of the most populous vessel in the fleet
 
                     }
                 }
